+++ src/rom/rom_data.cpp
@@ -2,17 +2,17 @@
 
 namespace spintool::rom
 {
-	void ROMData::SetROMData(const size_t start_offset, size_t end_offset)
+	void ROMData::SetROMData(const Uint32 start_offset, Uint32 end_offset)
 	{
 		rom_offset = start_offset;
 		real_size = (end_offset - start_offset);
 		rom_offset_end = start_offset + real_size;
 	}
 
-	void ROMData::SetROMData(const Uint8* start_data_ptr, const Uint8* end_data_ptr, const size_t start_offset)
+	void ROMData::SetROMData(const Uint8* start_data_ptr, const Uint8* end_data_ptr, const Uint32 start_offset)
 	{
 		rom_offset = start_offset;
-		real_size = (end_data_ptr - start_data_ptr);
+		real_size = static_cast<Uint32>(end_data_ptr - start_data_ptr);
 		rom_offset_end = start_offset + real_size ;
 	}
